proxy: Fix issue with incomplete body
When directly coping the body to the ResponseWriter it could happen that
the copy operation would fail with errors such as:

Error: write tcp 127.0.0.1:8080->127.0.0.1:43322: write: connection reset by peer

Or

Error: write tcp 127.0.0.1:8080->127.0.0.1:39806: write: broken pipe

Some tests showed that the reading was fine but the writing to the
ResponseWriter somehow failed. When buffering the response this doesn't
seem to happen.